requirements.txt
absl-py==1.4.0
altair==5.2.0
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
flatbuffers==24.3.25
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.42
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.12.1
idna==3.4
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.7.0
Keras-Preprocessing==1.1.2
libclang==15.0.6.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
narwhals==1.15.1
numpy==1.26.3
opt_einsum==3.3.0
optree==0.13.1
packaging==23.2
pandas==1.5.3
pillow==10.2.0
pip==23.0.1
protobuf==4.22.0
pyarrow==15.0.1
pydeck== 0.8.1b0
Pygments==2.14.0
python-dateutil==2.9.0.post0
pytz==2022.7.1
referencing==0.33.0
requests==2.28.2
rich==13.7.1
rpds-py==0.18.0
setuptools==67.4.0
six==1.16.0
smmap==5.0.1
streamlit==1.29.0
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.0
tensorflow-estimator==2.14.0
tensorflow_intel==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
toml==0.10.2
tornado==6.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.14
watchdog==4.0.0
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.14.1